refactor: UI 로직 분리를 위해 View 패키지 생성, 출력 코드 옮김
